Make image optional in tag search
I want to reuse this elsewhere
