+++ app/views/galleries/tag_searches/_tag_search.html.erb
@@ -1,7 +1,7 @@
 <h3 class="text-lg mb-3">Add tag</h3>
 
 <%= render(
-  partial: "galleries/images/tag_searches/form",
+  partial: "galleries/tag_searches/form",
   locals: {
     tag_search:,
     gallery: @gallery,
@@ -10,7 +10,7 @@
 ) %>
 
 <%= render(
-  partial: "galleries/images/tag_searches/results",
+  partial: "galleries/tag_searches/results",
   locals: {
     tag_search:,
     gallery: @gallery,
