+++ config/routes.rb
@@ -33,9 +33,9 @@ Rails.application.routes.draw do
   resources :galleries do
     scope module: :galleries do
       resources :images, except: %i[new create] do
-        resource :tag_search, only: %i[show], module: :images
         resources :tags, only: %i[create destroy], module: :images
       end
+      resource :tag_search, only: %i[show]
       resource :bulk_upload, only: %i[new create]
       resources :tags
     end
