+++ spec/system/galleries/image_tags_spec.rb
